+++ wayland-backend/src/test/many_args.rs
@@ -36,7 +36,7 @@ macro_rules! serverdata_impls {
                 None
             }
 
-            fn destroyed(&self, _: $server_backend::ClientId, _: $server_backend::ObjectId) {}
+            fn destroyed(&self, _: &mut (), _: $server_backend::ClientId, _: $server_backend::ObjectId) {}
         }
 
         impl $server_backend::GlobalHandler<()> for ServerData {