+++ wayland-backend/src/test/mod.rs
@@ -153,7 +153,7 @@ impl<D> server_rs::ObjectData<D> for DoNothingData {
         None
     }
 
-    fn destroyed(&self, _: server_rs::ClientId, _: server_rs::ObjectId) {}
+    fn destroyed(&self, _: &mut D, _: server_rs::ClientId, _: server_rs::ObjectId) {}
 }
 
 impl<D> server_sys::ObjectData<D> for DoNothingData {
@@ -167,7 +167,7 @@ impl<D> server_sys::ObjectData<D> for DoNothingData {
         None
     }
 
-    fn destroyed(&self, _: server_sys::ClientId, _: server_sys::ObjectId) {}
+    fn destroyed(&self, _: &mut D, _: server_sys::ClientId, _: server_sys::ObjectId) {}
 }
 
 // Client Object Data